improve script with random name package and format entries with count
