+++ README.md
@@ -1,3 +1,9 @@
-# NakedCPHstrikeraffle-master
+# NakedCPHstrikeraffle
 
 Only change the items with a #
+
+To run
+
+`pip install -r requirements.txt`
+
+`python raffle.py`
